Audio mixer cleanup

+++ LumoraGodot/Source/Godot/Audio/AudioBus.cs
@@ -28,7 +28,7 @@ public partial class AudioBus : IAudioBus
         _isValid = false;
     }
 
-    public float Channels => IsValid ? AudioServer.GetBusChannels(_busId) : 0f;
+    public int Channels => IsValid ? AudioServer.GetBusChannels(_busId) : 0;
 
     public List<IAudioEffect> Effects => _effects;
 
@@ -46,15 +46,17 @@ public partial class AudioBus : IAudioBus
         }
     }
 
-    public bool Mute {
+    public bool Mute
+    {
         get
         {
             EnsureValid();
             return AudioServer.IsBusMute(_busId);
         }
-        set {
+        set
+        {
             EnsureValid();
-            AudioServer.SetBusMute(_busId,value);
+            AudioServer.SetBusMute(_busId, value);
         }
     }
     IAudioBus? IAudioBus.Target
